// apps/web/lib/types/domain.ts
export enum BeltLevel {
  White = 'white',
  Gray = 'gray',
  Yellow = 'yellow',
  Orange = 'orange',
  Green = 'green',
  Blue = 'blue',
  Purple = 'purple',
  Brown = 'brown',
  Black = 'black',
}